refactor: Split auction tests
